move Form types from steplist/types.js -> form-types.js
