+++ protocol-designer/src/containers/ConnectedStepEditForm.js
@@ -13,7 +13,7 @@ import {
 import {selectors as labwareIngredSelectors} from '../labware-ingred/reducers'
 import {selectors} from '../steplist/reducers' // TODO use steplist/index.js
 import {selectors as fileDataSelectors} from '../file-data'
-import type {FormData} from '../steplist/types'
+import type {FormData} from '../form-types'
 import type {BaseState, ThunkDispatch} from '../types'
 
 import StepEditForm, {type Props as StepEditFormProps} from '../components/StepEditForm'
