+++ protocol-designer/src/containers/ConnectedStepList.js
@@ -4,7 +4,8 @@ import {connect} from 'react-redux'
 import type {BaseState, ThunkDispatch} from '../types'
 
 import {selectors} from '../steplist/reducers'
-import type {StepIdType, SubstepIdentifier} from '../steplist/types'
+import type {SubstepIdentifier} from '../steplist/types'
+import type {StepIdType} from '../form-types'
 import {hoverOnSubstep, selectStep, hoverOnStep, toggleStepCollapsed} from '../steplist/actions'
 import * as substepSelectors from '../top-selectors/substeps'
 import {selectors as fileDataSelectors} from '../file-data'
