+++ protocol-designer/src/steplist/actions.js
@@ -3,7 +3,8 @@ import type {Dispatch} from 'redux'
 
 import {selectors} from './reducers'
 import {END_STEP} from './types'
-import type {StepType, StepIdType, FormSectionNames, FormModalFields, SubstepIdentifier} from './types'
+import type {SubstepIdentifier, FormSectionNames} from './types'
+import type {StepType, StepIdType, FormModalFields} from '../form-types'
 import type {GetState, ThunkAction, ThunkDispatch} from '../types'
 
 type EndStepId = typeof END_STEP
